ubuntu-seedance-v2/seedance-requirements.txt
gradio==5.42
requests==2.31.0
pillow==10.0.1
python-dotenv==1.0.0